Experiment Drive Cycle with 'for' Event
Added the functionality of simulating drive cycles in Experiment class along with 'for' Event.
If the time specified in 'for event is less than the drive cycle time, the drive cycle is stoped at the time specified in 'for' event.
Else If the time specified in 'for event is greater than the drive cycle time, the drive cycle first extended for 1 week and then stoped at the time specified in 'for' event.